\name{Xrates}
\alias{Xrates}
\docType{data}
\title{
Exchange rate table
}
\description{
The is a table of exchanges rates to USD
}
\usage{data(Xrates)}
\format{
  A data frame with 184 observations on the following 2 variables.
  \describe{
    \item{\code{From.Cur}}{a factor with levels \code{AED} \code{AFA} \code{ALL} \code{AMD} \code{ANG} \code{AOA} \code{ARS} \code{ATS} \code{AUD} \code{AWG} \code{AZM} \code{BAM} \code{BBD} \code{BDT} \code{BEF} \code{BGL} \code{BGN} \code{BHD} \code{BIF} \code{BMD} \code{BND} \code{BOB} \code{BRL} \code{BSD} \code{BTN} \code{BWP} \code{BYR} \code{BZD} \code{CAD} \code{CDF} \code{CHF} \code{CLF} \code{CLP} \code{CNY} \code{COP} \code{CRC} \code{CSD} \code{CUP} \code{CVE} \code{CYP} \code{CZK} \code{DEM} \code{DJF} \code{DKK} \code{DOP} \code{DZD} \code{EEK} \code{EGP} \code{ESP} \code{ETB} \code{EUR} \code{FIM} \code{FJD} \code{FKP} \code{FRF} \code{GBP} \code{GEL} \code{GHC} \code{GIP} \code{GMD} \code{GNF} \code{GRD} \code{GTQ} \code{GYD} \code{HKD} \code{HNL} \code{HRK} \code{HTG} \code{HUF} \code{IDR} \code{IEP} \code{ILS} \code{INR} \code{IQD} \code{IRR} \code{ISK} \code{ITL} \code{JMD} \code{JOD} \code{JPY} \code{KES} \code{KGS} \code{KHR} \code{KMF} \code{KPW} \code{KRW} \code{KWD} \code{KYD} \code{KZT} \code{LAK} \code{LBP} \code{LKR} \code{LRD} \code{LSL} \code{LTL} \code{LUF} \code{LVL} \code{LYD} \code{MAD} \code{MDL} \code{MGA} \code{MGF} \code{MKD} \code{MMK} \code{MNT} \code{MOP} \code{MRO} \code{MTL} \code{MUR} \code{MVR} \code{MWK} \code{MXN} \code{MYR} \code{MZM} \code{NAD} \code{NGN} \code{NIO} \code{NLG} \code{NOK} \code{NPR} \code{NZD} \code{OMR} \code{PAB} \code{PEN} \code{PGK} \code{PHP} \code{PKR} \code{PLN} \code{PTE} \code{PYG} \code{QAR} \code{ROL} \code{RON} \code{RUB} \code{RWF} \code{SAR} \code{SBD} \code{SCR} \code{SDD} \code{SDR} \code{SEK} \code{SGD} \code{SHP} \code{SIT} \code{SKK} \code{SLL} \code{SOS} \code{SRD} \code{SRG} \code{STD} \code{SUM} \code{SVC} \code{SYP} \code{SZL} \code{THB} \code{TND} \code{TOP} \code{TRL} \code{TRY} \code{TTD} \code{TWD} \code{TZS} \code{UAH} \code{UGX} \code{USD} \code{UYU} \code{UZS} \code{VEB} \code{VEF} \code{VND} \code{VUV} \code{WST} \code{XAF} \code{XAG} \code{XAU} \code{XCD} \code{XOF} \code{XPD} \code{XPF} \code{XPT} \code{YER} \code{YUM} \code{ZAR} \code{ZMK} \code{ZWD}}
    \item{\code{Exch.Rate}}{a numeric vector}
  }
}
\details{
tmp
}
\source{
tmp
}
\references{
tmp
}
\examples{
data(Xrates)
## maybe str(Xrates) ; plot(Xrates) ...
}
\keyword{datasets}
